GUI with PyGame and PyGameMenu
No major change regarding the project. New file generate_GUI.py contains the code for generate.py as well as a UI layer for the menu to run and play the generated music.
Additionally, added a test_output.mid file that is the result of a generated sequence, which can be played to see an example of what generate.py sounds like.